fix api invoice order

+++ database/migrations/2025_06_06_082710_create_order_items_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string("item_name");
             $table->decimal("price", 20, 2);
             $table->decimal("item_price", 20, 2);
-            $table->integer("discount")->default(0);
+            $table->decimal("discount", 20, 2)->default(0);
             $table->enum('item_for', ['sale', 'sample', 'free'])->default('sale');
             $table->integer("quantity");
             $table->boolean("is_deleted")->default(0);
